requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
pymongo==4.6.1
paho-mqtt==1.6.1
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3
